requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
Pillow>=10.0
django-allauth>=0.63
PyJWT>=2.10.1
requests>=2.32.5
jwt>=1.4.0
